# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "POS / POS Referral Onboarding API",
    "description": "Backend for the digital onboarding flow (link generation, form, documents, approvals). Login/auth is provided by an existing system and is not part of this service.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://grid.leadsgenerations.in/recrument-analysis",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local dev"
    }
  ],
  "paths": {
    "/api/v1/links/generate": {
      "post": {
        "tags": [
          "links"
        ],
        "summary": "Generate Link",
        "description": "Called by the SM (Sales Manager) app after the SM logs in and enters\nthe POS / Referral name + mobile number. Creates the onboarding\nrecord and returns a token + shareable URL - nothing is rendered by\nthis backend; the URL points at a route the frontend team owns.",
        "operationId": "generate_link_api_v1_links_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateLinkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/links/{token}": {
      "get": {
        "tags": [
          "links"
        ],
        "summary": "Resolve Link",
        "description": "The frontend's onboarding page calls this on load (using the :token\nfrom its own route) to validate the link and get the prefill data\n(name, number, user_type, current status) needed to render the\ncorrect form/step. This is how the shared link \"renders\" on the\nfrontend - the backend never serves HTML, it only backs the route.",
        "operationId": "resolve_link_api_v1_links__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkPrefillResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/onboarding/{token}": {
      "get": {
        "tags": [
          "onboarding"
        ],
        "summary": "Get Onboarding Record",
        "operationId": "get_onboarding_record_api_v1_onboarding__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingRecordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/onboarding/{token}/form": {
      "put": {
        "tags": [
          "onboarding"
        ],
        "summary": "Save Form",
        "operationId": "save_form_api_v1_onboarding__token__form_put",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingFormUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingRecordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/onboarding/{token}/terms-and-conditions": {
      "put": {
        "tags": [
          "onboarding"
        ],
        "summary": "Update Terms And Conditions",
        "description": "Edits ONLY terms_and_conditions, independent of the general form-save\nendpoint. terms_and_conditions is set to a fixed value at record\ncreation - this is the sole way to change it afterward.",
        "operationId": "update_terms_and_conditions_api_v1_onboarding__token__terms_and_conditions_put",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TermsAndConditionsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingRecordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/onboarding/{token}/submit": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Submit Form",
        "description": "Final \"Review & Submit\" step. Runs basic form/document validation and\na duplicate check, then moves the record into the approval workflow\n(UNDER_REVIEW), matching the \"System Processing\" stage in the flow.",
        "operationId": "submit_form_api_v1_onboarding__token__submit_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/onboarding/{token}/documents": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Upload Document",
        "operationId": "upload_document_api_v1_onboarding__token__documents_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          },
          {
            "name": "doc_type",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DocumentType"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_onboarding__token__documents_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/onboarding/{token}/documents/pan-details": {
      "put": {
        "tags": [
          "documents"
        ],
        "summary": "Save Pan Details",
        "operationId": "save_pan_details_api_v1_onboarding__token__documents_pan_details_put",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PanDetailsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/onboarding/{token}/documents/aadhaar-details": {
      "put": {
        "tags": [
          "documents"
        ],
        "summary": "Save Aadhaar Details",
        "operationId": "save_aadhaar_details_api_v1_onboarding__token__documents_aadhaar_details_put",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AadhaarDetailsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/onboarding/{token}/documents/education-details": {
      "put": {
        "tags": [
          "documents"
        ],
        "summary": "Save Education Details",
        "operationId": "save_education_details_api_v1_onboarding__token__documents_education_details_put",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EducationDetailsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_api_v1_onboarding__token__documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_onboarding__token__documents_post"
      },
...
      "DocumentType": {
        "type": "string",
        "enum": [
          "PAN_CARD",
          "AADHAAR",
          "EDUCATIONAL_QUALIFICATION",
          "PHOTOGRAPH",
          "CANCELLED_CHEQUE"
        ],
        "title": "DocumentType",
        "description": "Which users_data file column an uploaded file is stored against."
      },
      "DocumentUploadResponse": {
        "properties": {
          "doc_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "doc_type",
          "url"
        ],
        "title": "DocumentUploadResponse"
      },
...
      "GenerateLinkRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150,
            "title": "Name"
          },
          "number": {
            "type": "string",
            "maxLength": 15,
            "title": "Number"
          },
          "user_type": {
            "$ref": "#/components/schemas/UserType"
          },
          "raised_by": {
            "type": "string",
            "maxLength": 50,
            "title": "Raised By",
            "description": "SM employee code"
          }
        },
        "type": "object",
        "required": [
          "name",
          "number",
          "user_type",
          "raised_by"
        ],
        "title": "GenerateLinkRequest"
      },
      "GenerateLinkResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "onboarding_url": {
            "type": "string",
            "title": "Onboarding Url"
          },
          "share": {
            "$ref": "#/components/schemas/ShareLinks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "token",
          "onboarding_url",
          "share"
        ],
        "title": "GenerateLinkResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LinkPrefillResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "number": {
            "type": "string",
            "title": "Number"
          },
          "user_type": {
            "$ref": "#/components/schemas/UserType"
          },
          "status": {
            "$ref": "#/components/schemas/OnboardingStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "number",
          "user_type",
          "status"
        ],
        "title": "LinkPrefillResponse"
      },
      "OnboardingFormUpdate": {
        "properties": {
          "dob": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dob"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "pincode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pincode"
          },
          "consent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConsentStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "OnboardingFormUpdate",
        "description": "Partial/upsert payload for the fields users_data actually has.\nterms_and_conditions is NOT editable here - it's a fixed value set at\ncreation, editable only via PUT .../terms-and-conditions."
      },
...
      "OnboardingRecordResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "number": {
            "type": "string",
            "title": "Number"
          },
          "user_type": {
            "$ref": "#/components/schemas/UserType"
          },
          "status": {
            "$ref": "#/components/schemas/OnboardingStatus"
          },
          "raised_by": {
            "type": "string",
            "title": "Raised By"
          },
          "dob": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dob"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "pincode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pincode"
          },
          "pan_card": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pan Card"
          },
          "educational_qualification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Educational Qualification"
          },
          "aadhaar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aadhaar"
          },
          "photograph": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photograph"
          },
          "cancelled_cheque": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled Cheque"
          },
          "terms_and_conditions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms And Conditions"
          },
          "consent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConsentStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "number",
          "user_type",
          "status",
          "raised_by",
          "created_at",
          "updated_at"
        ],
        "title": "OnboardingRecordResponse"
      },
...
      "PanDetailsUpdate": {
        "properties": {
          "pan_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pan Number"
          },
          "pan_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pan Name"
          },
          "pan_father_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pan Father Name"
          },
          "pan_dob": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pan Dob"
          },
          "pan_gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pan Gender"
          }
        },
        "type": "object",
        "title": "PanDetailsUpdate"
      },
...
      "ShareLinks": {
        "properties": {
          "copy_link": {
            "type": "string",
            "title": "Copy Link"
          },
          "whatsapp": {
            "type": "string",
            "title": "Whatsapp"
          },
          "sms": {
            "type": "string",
            "title": "Sms"
          }
        },
        "type": "object",
        "required": [
          "copy_link",
          "whatsapp",
          "sms"
        ],
        "title": "ShareLinks"
      },
      "SubmitResponse": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/OnboardingStatus"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "SubmitResponse"
      },
      "TermsAndConditionsUpdate": {
        "properties": {
          "terms_and_conditions": {
            "type": "string",
            "title": "Terms And Conditions"
          }
        },
        "type": "object",
        "required": [
          "terms_and_conditions"
        ],
        "title": "TermsAndConditionsUpdate"
      },
      "UserType": {
        "type": "string",
        "enum": [
          "POS",
          "POS_REFERRAL"
        ],
        "title": "UserType"
      },
...
    }
  }
}